% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator0null.R
\name{beta1estimator0}
\alias{beta1estimator0}
\title{beta1estimator0}
\usage{
beta1estimator0(X1, X2, sigma, n, nbSimu)
}
\arguments{
\item{X1}{1st data column}

\item{X2}{2nd data column}

\item{sigma}{sd}

\item{n}{data length}

\item{nbSimu}{number of simulations}
}
\value{
A vector of beta1 simulated
}
\description{
simulations for beta1
}
