% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator0null.R
\name{densitybeta1}
\alias{densitybeta1}
\title{densitybeta1}
\usage{
densitybeta1(X1, X2, sigma, p, h = 0.01, max = 2)
}
\arguments{
\item{X1}{1st data column}

\item{X2}{2nd data column}

\item{sigma}{sd}

\item{p}{the 4 probabilities obtained with the probaEstimation function}

\item{h}{step size}

\item{max}{max value}
}
\value{
x y and d0 density at 0
}
\description{
density of the estimator for beta1
}
