% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator0null.R
\name{probaEstimation}
\alias{probaEstimation}
\title{probaEstimation}
\usage{
probaEstimation(X1, X2, sigma, n, nbSimu)
}
\arguments{
\item{X1}{1st data column}

\item{X2}{2nd data column}

\item{sigma}{sd}

\item{n}{data length}

\item{nbSimu}{number of simulations}
}
\value{
a vector of 4 estimated probabilities
}
\description{
Estimation of 4 probabilities
}
